cleared debug msg; reorganized files
