+++ main/forms.py
@@ -32,6 +32,7 @@ from .models import (
     Membership,
     User,
     Address,
+    State,
 )
 from .choices import STATES
 from django.contrib.gis.db import models
@@ -287,3 +288,18 @@ class RepresentableLoginForm(LoginForm):
             del field.widget.attrs["placeholder"]
         self.fields["login"].label = "E-mail"
         del self.fields["login"].widget.attrs["autofocus"]
+
+class SubmissionAddDrive(forms.Form):
+    def upd(self, state):
+        """
+        Req: state must be a valid, uppercased, state abbreviation
+
+        Creates a dropdown of drives that satisfy the rules:
+            * Are in state
+            * Are active
+            * Are not private
+        """
+        all_drives = State.objects.get(abbr=state).get_drives()
+        drives_to_add = [d for d in all_drives if d.state==state and d.is_active==True and d.private==False]
+        choices = [(str(d.id), str(d.name) + ' - ' + str(d.organization)) for d in drives_to_add]
+        self.fields['Add a new drive'] = forms.ChoiceField(choices=choices)
