+++ main/templates/main/submission.html
@@ -28,8 +28,6 @@
 {% endblock %}
 {% block content %}
 <div class="container-fluid w-100">
-  <pre>{{ form_info }}</pre>
-
   <div class="row row-wide light-green-bg pb-5">
     <div>
       <script type="text/javascript">
@@ -174,7 +172,7 @@
                     <button type="submit" class="btn btn-primary">Submit</button>
                   </form>
                 {% else %}
-                  <p class="header-text small">Your community map has already been submitted to an organization in your state</p>
+                  <p class="header-text small">This community map has already been submitted to an organization</p>
                 {% endif %}
               </div>
             </div>
